feat: change gallery from using flex to grid, change the sizing of the image inside the gallery
Flex was used for gallery, but when configured to stay in the center, the last image would place itself also in the middle, for example when there are 3 columns, it would place itself under 2nd column, if it was alone in the last row. This was changed, so that the image appears at the beginning of the row. The image was taking only a specific amout of width, this caused the image to take too little place and appear small on mobile devices.