feat: configure images to load smaller versions and set various format, based on availability
Some images were loaded of big size, even if a much smaller version was displayed, therefore I set a specific size to load 350x500 which is a bit more than the biggest picture on the site would have. It is still much smaller than before when images of size 1000x1050 could be loaded. I configured images to be loaded with different formats based on availability, first avif, then webp and then png. It is done in order to make images loading size smaller.